lambdas/process_post_ideas/requirements.txt
requests
openai
psycopg2-binary